requirements.txt
streamlit==1.39.0
pymongo==4.10.1
python-dotenv==1.0.1
bcrypt==4.2.0